+++ lib/community/widgets/post_card_metadata.dart
@@ -1,3 +1,4 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -7,6 +8,7 @@ import 'package:thunder/core/auth/bloc/auth_bloc.dart';
 import 'package:thunder/core/enums/full_name_separator.dart';
 import 'package:thunder/feed/feed.dart';
 import 'package:thunder/shared/avatars/community_avatar.dart';
+import 'package:thunder/shared/avatars/user_avatar.dart';
 import 'package:thunder/shared/icon_text.dart';
 import 'package:thunder/shared/text/scalable_text.dart';
 import 'package:thunder/thunder/bloc/thunder_bloc.dart';
@@ -207,13 +209,22 @@ class PostCommunityAndAuthor extends StatelessWidget {
               ),
               onTap: () => navigateToFeedPage(context, communityId: postView.community.id, feedType: FeedType.community),
             ),
+          if (state.showPostCreatorAvatar)
+            GestureDetector(
+              onTap: (compactMode && !state.tappableAuthorCommunity) ? null : () => navigateToUserPage(context, userId: postView.creator.id),
+              child: UserAvatar(
+                person: postView.creator,
+                radius: 16,
+              ),
+            ),
+          if (state.showPostCreatorAvatar) const SizedBox(width: 8),
           Expanded(
             child: Padding(
               padding: const EdgeInsets.only(right: 6.0),
               child: Wrap(
                 direction: Axis.horizontal,
                 alignment: WrapAlignment.start,
-                crossAxisAlignment: WrapCrossAlignment.end,
+                crossAxisAlignment: WrapCrossAlignment.center,
                 children: [
                   if (state.showPostAuthor || communityMode)
                     Row(
@@ -263,7 +274,7 @@ class PostCommunityAndAuthor extends StatelessWidget {
                           ),
                       ],
                     ),
-                  ),
+                  )
                 ],
               ),
             ),