+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Added support for displaying comment origin instance - contribution from @ggichure.
 - Added ability to favorite communities on the drawer
 - Added ability to set keywords for filtering post titles/body
+- Added setting to show/hide user avatars in comments and posts - contribution from @ggichure
 
 ## Changed
 - Added new items to the Post and Comment actions sheet - contribution from @micahmo